EZP-31651: Move and mark classes as deprecated
